Make Evolution more resilient against concurrent modification exceptions.
BenchedMoves is quite vulnerable to concurrent modifications as it does no internal locking. The code changed in this commit showcases that issue extremely well: As `pokemon.benchedMoves` was updated one-by-one and each update send a separate BenchedMovesUpdatePacket which would all be handled in parallel by the Client evolving a Pokémon which lears a bunch of moves on evolution was an almost sure-fire way to crash the game.

This *isn't a fix* for that! It's merely a hotpatch to make sure this common a crash doesn't sneak itself into the next release.

Fixes: #1536